fix(CarNames): CarNames 추가로, 테스트코드 수정
